[ALPS05037435] GPS: Sync code from Android Q to R
Sync code from Android Q to R:
| * 0054dec (origin/alps-mp-q0.mp2)
            [ALPS05015487] GPS: Add debug for too much RX data issue
| * 845a88f [ALPS05002403] GPS: Add more polling of DSP reg for debug
| * fee419d [ALPS05002403] GPS: Add MT6885 deep stop mode support
| * c0e81a3 [ALPS04997182] GPS:fix coverity use of uninitialized scalar
| * 8eba9f8 [ALPS04989423] GPS:change margaux emi mpu set region
                                protection api
| * 91bf5ba [ALPS04961016] GPS: Transfer suspend mode to FW
| * fb2a09a [ALPS04957036] GPS: Add protection for dsp debug via MCUB
| * 986f23c [ALPS04954124] GPS: Keep conninfra bus clock when doing GPS on
| * 176cc4f [ALPS04956303] GPS: Fix corner case of stopping dma
| * 353468b [ALPS04952271] GPS: Handle the corner case of taking
                                HW semaphore

MTK-Commit-Id: fa73ebcd4ef9e8a133037b2132e0410d1cc3ba7c

CR-Id: ALPS05037435
Feature: Location Chipset Capability

+++ data_link/lib/inc/gps_dl_time_tick.h
@@ -16,7 +16,7 @@
 #define GPS_DL_RW_NO_TIMEOUT (-1)
 void gps_dl_wait_us(unsigned int us);
 #define GDL_WAIT_US(Usec) gps_dl_wait_us(Usec)
-unsigned int gps_dl_tick_get(void);
+unsigned long gps_dl_tick_get(void);
 int gps_dl_tick_delta_to_usec(unsigned int tick0, unsigned int tick1);
 
 #endif /* _GPS_DL_TIME_TICK_H */
